Finalised BMS  schematic befeore testing PCB
is rough design without considering guitar dimension, little EMI considerations. not all concepts in design have been ensured to work properly. will test each sections viability before testing entire Circuit.

first design with the batterys and the circuit on the same board. initially had the idea of a 3d printed holder for the cells and a board on top.
 then 3 layers of single sided pcbs triple deckered with smd battery mounts. finally to this style.
although the board is larger in width and length, the height is reduced which should permit a slot in battery design for the guitar rather than a large module with a rail on the side of the guitar for battery to connect to.
This should make the design simpler and will look nicer becasue there will only be a small plastic plate exposed, rather than a chunk of the guitar body being plastic, ruining the over all finish of the guitar.