updated bms schem and pcb and added initial schem for charger
fixed charging switch logic for bms board, updated jst conns and now instead uses the supply v as the sw V+ to turn off the mosfet to then charge. should work considering it takes 100ns ~ to switch. added caps in schem but still havent calculated their values, nor are they on the pcb.

created the initial schematic from my notes for a charger that works in respect to the BMS board, is in the same cad file but different sheet to bms and is a seperate pcb.

still need to finalise the caps and also determine pcb via diameters with what i have available and the size requirements. will follow the suggestion for via size from kicad forum