upgraded interface software
improved controls of audio pages. streamlined more page processes for less code

added SD card utilisation which reads a txt file on the card and sets all initial variables there stored values. continuing to work through updating values in the txt file and experementing with the method for uploading the data.

currently the delay fx updates the file every time the delay time is increased which causes a delay (pun not intended) in the changing of the value.

the other 3 FX use a still not working method that stores up to 3 different variables changed and their new values. if its changed again the same variable will be changed. once 3 is reached each of them are added to the file.
 - this has many bugs and doesnt work most of the time as intended so will likely be using what i learnt from these experiments and making it a time counter based process instead that will save all the values at once rather than only the ones that have been modified, a sort of blanket update of the variables. this may not be as efficient but seems to be the most achievable solution. will improve upon this in future variations.

next technology after this will be adafruit neopixels in a chain on a seperate data pin for aestetic lighting of the guitar. will be the last hardware addition to the interface.

after the led's are added in their simplest form and the distorted sound issues with the amplifier and speaker are resolved, a circuit board and enclosure will be constructed to house and continue testing the platform and improving the code for it. the circuit will realistically not be improved and the enclosure will give insight on designing and modelling the future mounted housing unit for the electronics when they go inside the guitar.

after the SDsection, the audio section and a basic lights on/off part for the leds, the interface process will be put on hold and i will return to the battery, motivated from this quite speedy development of the interface system.

after the batttery circuit works and an adequate enclosure and mechanism for attaching the battery to the guitar is completed, the last step will be the design and manuafacturing of the guitar body itself, the last big design challange...