Updated SD Card Usage and designed PCB for AFPI Device
Created a new method that reads all existing variables and updates all existing variables instead of only the ones that are updated to update all variables in the file. currently this is only actioned when the home button is pressed, however a regular interval update will be put in place to have the best chance of saving the values.

the next thing to implement for the sd card is to create a temporary file to store the variables while writing it so that it can prevent corruption causing a loss of data so that there will always be a backup file.

the AFPI Device V1.0 PCB is introduced as well to incorperate all the hardware into one circuit board, including the teensy board and audio adapter, connectors and amplifier. is currently a single sided board with not much thought put into the connector placement, however i will make another version thats potentially double sided and with consideration on putting the connectors that will not leave the AFPI's enclosure on top (for the OLED, Rotary encoders and switches), and for the connectors that leave the enclosure, they will be mounted on the under side for ease of connection. and to minimise enclosure side and there wont need to be room around the PCB for wires (RGB LEDS, guitar line in, line outs, power in etc)

the adafruit quad rotary encoder board also stopped working on the date of the last commit, which was frustrating as it limited my progression, however will hopefully get here this coming week.

also ordered connectors for the AFPI board and Active Pickups which should come in a week.

will ideally hope for the AFPI Board and enclosure to be finallised within the next 3-4 weeks