restarted AFPI code development
now with the device built, most of the features need their code written to test.

have commented all serial comms to work without the connection and added the ability to take the line in input and mix it with a aux input, however the aux input audio quality is wierdly scrambled and it seems the higher frequencies are being lost

have begun testing an audio output switch to switch between the speaker amp, headphone out or aux out, and while the headphone and i believe the aux out will mute and unmute when not in use, the speaker amp doesnt output anything.

testing the circuit the op amp recieves the 9 volts and when turning the gain and volume trim pots to the max noise can be heard, but its just noise and no signal makes an difference on the output speakers, the only thing that can be heard is the interference caused by the sd card when its being written to when the home button is pressed, which is heard by the other outputs as well.

testing the output on the audio shield there is a voltage on the output, however i am not sure if the voltage is correct at about 0.5-0.7v. will need to first figure out what voltage should be correct for an out put and whats enough for the circuit to use. after that i will need to then craft a audio probe and use the osciloscope to measure the input and output on the circuit to determine if its a physical problem and if so, figure out whats going wrong and come up with a solution.

if the circuit is fine and the voltage outputs are wrong or are correct but still no sound, then the issue most likely is a code problem and the code for allowing the output for the amp circuit and for controlling the gain/volume of the output will need to be revisted and solved to allow the speaker to work.

after the speaker circuit is fixed, the next thing to do will be to reformat the code to utilise librarys to clean up the main source code as its becoming increasingly more difficult to navigate, incorperate features and update functions due to its size of code and the amount of various unrelated functions and sections all jumbled up together, making programming it confusing and more time consuming. it was always expected that i would do this at some point but after returning to the code after making the AFPI, im having to reunderstand my code and learn how to navigate it which is not easy. i have successfully learnt how to set up a library however there is still alot to learn and understanding how to incorperate variables across the librarys functions, other library functions the source code functions will be the main challange.