Solved Speaker Issue
After reviewing the pcb layout and the schematic, and reviewing each component, i noticed that the pin outs on the symbol for the MPF102 on the schem didnt number the pins correctly to their functions. while the drain pin was wired correctly, the gate and source pins were reversed, causing it to not work properly. this mistake occured because i used a standard n channel JFET symbol and a generic TO-92 package during the design phase as kikad didnt have the MPF102 in its libraries, sourcing one was challenging and i just decided to use the generic symbol as it was the same.

While i did check for the different pinouts as i was aware they vary from models, with the symbol displaying the same outputs as the ruby amp schematic in the right positions i thought the numbering was correct, and the pin table for the symbol was numbered 1, 2 and 3, the same as the components pins left to right, i overlooked however the fact that the numbers didnt correlate with their correct functions. Once it came to PCB design i didnt think about it much as afterall, i had gone over the schematic muiltiple times and the outputs matched original schematic, i also did the electrical rules checker on the schematic and on the pcb after i designed it, and both were fine, as this was overlooked.

 i carried on assuming it was correct but i always expected their to be an issue in the circuit and was prepared for that fact and read analysis's of the circuit online to understand how it should operate so i was able to trace the circuits components to identify how it operated. When i realised the pins on the physical circuit were funny thats when i realised the mistake, luckily so early in the circuit so i didnt have to go far, although i did check voltages on the LM386 to ensure it was getting voltages.

initially was worried that their was no ground from the power supply reaching the circuit as i hadnt connected a direct ground source to the amplifier and instead used a ground plane with the ground only coming from the output of the teensy audio shield, assuming they should be connected, which was a rookie error looking back and shouldve realised it was correct, not to mention i couldve just used a multimeter to test the ground connection. this was my initial worry but after inspecting the power supplied to the op amp i realised it was completing the circuit with 9 volts so it was fine. this design choice most likely has impacted the audio quality as there is a bottle neck for the ground connection as the only path will be through the teensy shield and not the ground plane designed to reduce noise and interference. the pin may also be damaged as the current for powering the entire amplifier circuit will be going through the teensy shield as the gnd line will complete the circuit this way, potentially  exceeding the current limit of the pin. due to this i may incorperate a wire to connect the 2 ground planes together to mitigate this risk. i will also at some point update the schematic and pcb layout to incorperate the two ground planes for reference and for future creations.

for now though i a have already updated the schematic and altered the pcb layout to be correct, but i wont be recreating the board and instead will just twist the pins around and solder them. i initially considered making a small pcb on the cnc router for speed that would have header pins soldered between the main board and itself and then have a footprint with tht holes that were correctly wired to re attach the component correctly, however it was just simpler to solder a new one on with twisted legs and make use of heat shrink if neccesary.

after purchasing a new mpf102 and cut the old one and desoldered it away, the pad i mistakenly set to be the gate but was actually the source came off with the component without applying the soldering iron, even after taking my due diligence to carefully use the iron sparingly and at a low temperature (300c). it wasnt the end of the world however as i could still solder the source and drain pins to their appropriate holes and the gate pin was just bent to reach the smd Restistor 1m5 R1 as this was the only component in the circuit before the JFET. this wasnt too dificult and the pins are still adaquately spaced.

when it was tested it clearly worked straight away and while there is still a level of noise, adjusting the gain trim pot and incorperating noise filtering in the circuit will mitigate this. the main concern at least will be stopping noise when the speaker is not in use as althought the guitar tone doesnt come out when using a different output, there is still ground noise, most likely from both the amp circuit and definetly the teensy board and shield, as when the speaker is connected the sd card, when reading and writing will vocalise the noise picked up from the action of writing to the card, and even when the headphone jack is the primary output the speaker takes all this noise and is barely heard in the headphones compared to the speaker, and when not connected the noise is primarily head in the headphones, likely due to the speaker using more current and  potentially more resistance than the headphones. its not a massive concern compared to the other parts of the circuit as it only happens when the sd card is used, which is only when the home button is pressed, but when i incorperate a proper periodical updating of the card to ensure the values are saved more often without user input, this noise will be an issue as you dont want it to occur when playing. a crude solution would be a switch to turn off the speakers gnd conenction which would stop noise from the speaker when playing through other outputs, however this is not what i really want and i want an elagent and simple solution.

for now though it works and im stoked as the circuit finally works, i wont be remaking the board at least for this guitar, due to time constraints on the project considering the christmas day due date and i still need to make the battery and body istelf and code more all while doing uni and work and a social life, but for now its looking good. now i can actually start programming all the features properly, make a 3d printed insert for the jacks with an acrylic plate similar to the AFPI enclosure but on the guitars side, and then make the NeoPixel lighting boards mounts with frosted clear acrylic and program it to work. then i can focus on the next stages of the circuit.