First round of modularisation of interface code.
successfully incorperated the use of librarys in program to sperate different sections of the code, readability and will inevitably efficiency as the code was too long and complicated for one page. for now only the global variables and definitions, intitial setting up of program and the setup functions code have been removed, the loading screen display and sd card functionality with the sd card section being the only system to be fully modularised.

for now there is still an over use of global variables and when each section is fully and properly seperated i can then start encapsulating the sections in classes etc. All in all good progress considering my last attempt was a failure due to inexperience and i considered alot how to reformat the code before i started the fix. i expect that by the end of the week i will have seperated the sections and encapsulated the sections. then i can finally start battery voltage measuring and rgb control and its physical components, and then from there i can do any critical features that need to be done for the completion including power saving features. then stop coding and instead design, build, test and finalise a battery, circuits and enclosure along with the in out jack port insert which will have the power circuitry on board as well as the connections for the jacks on board and connecting to the AFPI. i can then go back to coding and try and implement as many features as i can before the end of uni becasue thats when i need to stop everything and start the preperation, design and building of the guitar body which needs to take first priority by the deadline of last uni exam.

i can code during this but only when there isnt something i could do for the body section. then hopefully with time to spare before christmas day i can code the rest of the features and hopefully a distortion effect, but only time will tell if i can meet this time line but so far, so good.