second round of improving current interface code
i have created all libraries neccesary for seperating sections of the program into their respective sections, for each of the page sections aswell as the peripherals code for initialising and sending and recieving inputs/outputs. the backend processing is also seperated for the audio processing. currently the code can still be improved as the config file holds variables that arent neccesary for all the systems and can be intergrated into their sections, not to mention the over use of global variables which will be improved. once the last round of code improvement is done i can begin incorperating new functions and updating previous code. i may have over used librarys, but it makes the code use so much easier to read. the reformatting and improvement of this code has actually been alot easier than i imagined it would be which was thanks to the research and prep i did before hand on techniques for modularising code for microcontrollers. i actually put off reformatting my code for a week or so cause it was just so messy (and cause i had assignements for uni) but its coming along well!