incorperation of new features to interface code
added the following new features:
- able to manipulate the OLED contrast to set the brightness of the screen in the settings section, currently between 0-255 in steps of 5 but will make it smaller in future for easier use
- added screen saver functionality to count 10 seconds and if 10 seconds is reached, the screen is cleared and the program waits in pergatory for a interaction with the buttons/encoders. if one is pressed the the timer is reset and the code is returned the previously used function. whenever an encoder is turned or pushed or the home button is pushed the timer resets to 0 and starts counting again. currently works mostly although the landing page only works with the #3 encoder which i havent found why, but for now its ok and will be improved later.
- last change made was to reverse the operation of the encoder rotation so now when turned clockwise it relates to increasing the variable value in question and if turned counter clock wise it relates to decreasing the value, rather than the other way around which was kept as that for way too long.

next features will be sorting out the volume and gain control between the 3 audio outputs to seamlessly transistion the control of the volume and gain control to match the current output, with a conversion rate for the different volume variables as they are set with very different functions.

after this the low power mode will be next and after that the sd card file will be updated to incorperate all these new variables.