added volume and gain control to specific outputs depending on chosen output
while still janky and variables arent very linear for the speaker and aux, i can control the vol and gain on the specific output chosen now, rather than just the headphone output, now whats left to finish the audio section is to initialise it on start up to know what output should be used when turned on and set the appropriate output and its vol and gain. then adding all these new variables for these controls to the sd storage for remembering the settings.

in the future the variables shown to the user will need to be more uniform as the vol for the speaker cuts out at 0.65 so that needs to be made 0 and 1.0 needs to be made max and then have linear steps between. but thats it for the volume and gain control for choosing specific outputs section.

a distortion page was also added aswell as code to test out a design found online, however it doesnt alter the output so i need reattempt later on. i have found a different source code for another design for a distortion effect that looks more understandable so will attempt that when the other more critical features are added due to my time constraints with my deadline and uni commitments.