Completed Audio Output Control and updated SD card storage system
Fixed all known bugs in audio output section, while the code for determining the current audio output and directing variable setting/carrying out gain/volume update is somewhat improper and isnt an elagent solution, it works as intended and correctly initialises the output and updates the correct values while seamlessly switching outputs.

the sd card variables were also updated for not only the audio output section but also any other new features that hadnt had their variables added. this was done by also using AI to create for me a program that when run, rewrites over the data file on the sd card with content stored in the code for it, so when adding new functions and their variables to the AFPI.

all i need to do is add the new variables to the new program as incorperate them and then once the feature is added to the main code and run the file updater program to update the stored variables, making testing and adding new functions easier.

otherwise would have had to take the top panel off the device to remove the sd card to then use a sd card to usb adapter on my computer to access the data file and update it, then putting it all back together and then test. to do this each time when adding or modifiying features would have been such a waste of time, now i can just spend 10 seconds rather than minutes each time. the plan is to add a debug feature to the device which will allow me to do this within the program and send the new data over serial to make it even more effective. the debug section will be a secret one that can only be accessed by doing a set button press sequence on startup, like a BIOS. it will also provide more indepth info on the code and also the stats on the CPU and RAM etc for testing purposes to improve its computing and power efficiency, however this is a future addition so no need to worry on this for now.

the next section i will be incorperating will be the RGB control which will require physical designing and testing aswell which will be one of the larger future sections to add, so stay tuned cause its coming.