introduced RGB control
I soldered together 6 ws2812 LED's and fixed up the rgb connector wiring so that it wasnt in the way of the usb programming cable. i then programmed the interval timer to check when 20ms has passed during the normal operation of the program and if so then the lights are set to their appropriate values and effects. this allows the lighting to be dynamic during other processes, meaning it doesnt have be 1 solid colour and instead we can have many different styles of lighting.

the initialisation of the leds was set up and incorperated as well as the neccesary variables added to the sd card storage system. from there i added a simple on off of a basic flowing effect that changes the hue of each light one at a time controlled by the encoder 4 on the rgb section when pressed. after this i added a basic user interface to the display to show the 4 parameters for the leds, rgb active, brightness, colour and effect. the last new thing for the rgb that i set up was the brightness controll between 5 and 255 in steps of 5 points. this actively changes the values and brightness of the leds dynamically and will be utilised for different effects in the future.

 but thats it for now, will add a basic couple of effects next and colours for the lighting and potentially a second page for modifying other parameters specifc to the different effects like rate etc. after the basic code for demonstrating the functionality of the leds the physical components will need to be designed and manuafactured and assembled for mounting and covering the leds within the guitar, then when the final guitar design is in motion and when its time to go back and update/improve and add features in the code including the rgb code structure and the available options for colours and effects.