completed min specs for lighting and Low power mode
The lighting code has been completed to the point where the only things to add is more lighting modes. the 2nd light mode atm will be turned into another colour option as its a slower rainbow than the current one, but for now ill leave as is so i can switch between modes, but will be changed when i return to the rgb code towards the end of the project when i add the rest of the other features for the other systems that arent mission critical features.

the low power mode feature has also been incorperated and for now it reduces the led brightness to max out at 50 instead of 255, the same goes for the oled contrast. the screen saver timer is also changed from 10 seconds to 5 when the mode is on. there are other features i wanna add to this low power mode to reduce power consumption such as led update interval, oled refresh rate and speaker volume and the teensy processor clock spead, however im not adding these for now as some may not actually improve the power use and i havent experimented with the oled refresh rate and teensy clock speed setting so i dont want to incorperate them just yet until the rest of the mission critical features are added and i have the rest of the features set up properly so i can test the impacts reducing these values cause.

but for now these 2 are now done for now and i can move on and do the battery voltage measurement portion to get the last mission critical feature out of the way. after this one is done ill likely add the equalisor and bass control features to the EQ page as this is expected to be simple as the functions are built into the teensy audio library. after this one ill move straight onto the battery design/manuafacturing and then the I/O jack enclosure with its built in pcb for holding the battery measuring resistors and the automatic switch circuit between the battery and 9v jack.

it is currently the 5/9/24 and i have 7 weeks until the end of week 13 of uni sem 2, meaning i have 7 weeks + the time until my last exam to do the tasks listed above before i have to move on to the design of the guitar body to make sure i dont run out of time before the deadline of christmas day. ill return to the code while im in the designing the body phase when i can afford to take breaks from the body and or when there isnt anything i can do in the mean time (i.e paint drying, delivery times, automated machine work not requiring human involvement etc). im getting there and i reckon im on time for meeting my schedule date.