Completed Lighting Board and 3d Model of lighting assembly
After finishing the minimum requirement code for the RGB lighting to function i have now designed the 2 PCBs that will house the LEDS as well as designed the mounting of these board for to the guitar along with its top plate.

the PCBS are numbered 1 and 2, 1 being placed between the neck and middle single coil pickups and 2 being mounted between the middle and bridge pickups. board one is the main PCB as the output from the AFPI will be connected into that board via a JST connector, then through a capacitor to smooth power fluctuations and a resistor to limit the strength and noise of the data signal to the leds. for testing the leds originally i utilised these 'puck' like modules which each individually contained the resistor and a capacitor as well a sthe ws2812B led itself on a single octagonal pcb, allowing easy wiring up for testing the system. for the actual pcb however these wont be used and these components will be soldered directly to the board individually.

the mounting holes are M2.5 due to size constraints and the boards are the same length as the pickups and are thin enough to fit in between a strat style pick up layout.

the boards are connected via a jst conn on each to carry the voltage, GND and data signal between the two and on the end of the 2nd board another jst connector is added incase i incorperate more leds in the design in the future, mainly if i choose to make my own neck with built in leds on the inlays i can just plug it in and not require any additional circuitry.

the clear acrylic top plate is bolted to the board through 10mm pcb spacers on each side of each board to have a standard distance between the plates for an even lighting. the top plates are designed to have the mounting areas 1.5mm high and the viewing area of the Lighting to be 3mm so that when mounted to the pick up mounting plate the pick up plate will have the same 1.5mm area removed on the underside and the viewing area removed so that the clear acrylic is flush to the main plate. the two clear plates are mounted to the main top plate by bolting the top plate through the clear plate and fastening them to the pcb spacers, ensuring reigidity. if needed to fill in any gaps between the 2 sheets of acrylic, silicon will be used to seal the gaps, which while this means the 2 plates wont be seperated it does still allow the bolts to be undone, disconnecting the lighting boards from the guitar itself, maintaining the ease of assembly and disassembly design rule i set early into the project to ease the incorperation of all systems/components into the guitar physical design, which was a big source of frustration when i first attempted this kind of project around october last year.

I will be making the acrylic plates as i am for all the acrylic plates, but this time ill use my knowledge of pcb making at home and demonstrate my ability by making the lighting boards, as they are relatively simple and i wish use the knowledge and abilitys i gained through the research phase of pcb design in this project and not let it just be theory. i have done many boards before but they were only Pedal circuits back when i was still pursuing the analog FX route.

Ill probably hold off on making these for the moment as uni is kicking up so ill wait on doing any manuafacturing until i can afford to spend long periods of time on the project, however ill still do designing for other parts of the project like the battery and I/O board that will act as the stopping off point between the AFPI and Battery as well as route the audio signals to the appropriate In and Out ports. im also not making the production of these rgb partsa top priority for the moment as there are no other systems that need these to be made to be able to continue, except for the final physical design, where as the battery and IO board are needed to complete the battery voltage measuring. once these last three main sections are manuafactured and are integrated into my testing of the whole project, i will be able to focus on the last big section of the project, being the Body design and manuafacturing.

but im getting there..