designed and awaiting testing of Battery Circuit Design and Enclosure functionality
i have fully simplified the batterys circuitry down to reduce the amount of time spent on the battery to be a little as possible to ensure the minimum requirements are met and that i ensure i meet the dead line. when i make the second battery which will be done after the full prototype is completed, ill create a new battery design/ charger/ internal circuitry to make the design of it better, mainly by making the batterys purpose to be the solely the vessel that houses the battery cells and the discharge circuitry (voltage regulating etc) will be housed in the internal sleave of the guitar that the battery is inserted in and then the charging circuitry will be kept within the charger, this will limit cost of multiple batteries as the charger circuitry will only need to be replicated for each charger rather than each battery, same with the discharge voltage, as there will be more batteries a person will use than they will have guitars and chargers (realistically 1 of each, although if they get more they would still benefit). this would also allow the batteries to be smaller, allowing for less needed area within the guitars body, allowing for less strict design rules for accomadating the battery, this would potentially allow for a 2x2 cell layout rather than the current 1x4 which was needed to allow the batt to fit within the guitar bodys depth as an 8mm thickness of wood on either side was required along with the internal sleeve and the batteries enclosure itself. the charger and disharger circuits will also ideally use custom made boards rather than bms boards from aliexpress which are overpowered for my needs and dont exactly fit my batteries specs, reducing their lifespan. i have been looking at 4s li ion charger ics like the MP26124 and that seems like a suitable option for a future design, however i wont be worrying about that until then. i forgot to also mention that i removed the use of a temperature NTC sensor as that circuitry would take too long to accomidate within the device, and as im well within the batteries use limits i decided that i wont need to include it, at least for this design as the v2 battery will for sure have it to meet the safety requirements of the battery. but for now according to the bms board description it has temperature protection built in to the board but doesnt say how it does this and what temp it cuts off at, and inspecting the board without any schematic i cant figure out how it does it, but something is better than nothing.

the batt enclosure is designed to fit each 4 cells snuggly with gaps between the front section, the cell mounting area and the back section which houses the circuitry for air flow throughout the battery, with a small vent on the front for air flow in which will be covered with dust protecting fan mesh. the lock mechanism works with a handle that is locked in to a spring loaded bolt that will lock in when inside the to be designed sleeve but and the handle must be lowered down 3mm and pulled out to remove the batt. the magnetic pins on the back will lock in to the pins that will be on the sleeve on the other side and due to the pins being on the left side and the rails on the base left and right of the battery enclosure, the batt will only be able to be inserted one way, ensuring no misconnections or shorts occur during use.

as of right now the enclosure is being printed and will be tested and potentially iterated more where necessary and the batt will be pretty much done, minus the charger circuit and enclosure which wont be hard, just copy the design for the sleeve and modify to fit the voltage boost/regulate circuit to power the charging bms board inside.