updated battery enclosure design to v1.1
after printing and testing the battery design i found multiple flaws in the design:

- the pockets for the spring bolt, handle insert to bolt and bolt hole on top plate were too small and needed to be widened with files and a dremel.
- the sleeve wasnt made big enough for a smooth slide and had to be 'broken in' and would get stuck and had to hammer out the battery.
- the battery wasnt easy to remove as there was no proper grip to pull it out with.

to fix these issues i resized the sleeve 0.7mm bigger on the sides and 0.5mm on the top for a smoother slide and the bolt was redesigned to more easily fit the handle into. 2 holes on either side of the spring handle were made at 20x15mm so a finger each side can grip the batt when removing it from the sleeve.

after reprinting the design it worked alot better, although the print for the main batt enclosure was a  bit poor on the right rail portion and needed limited filing to fit better, this isnt design related and only a tool issue.

the finger holes took up the air vent space, however being a prototype the batterys caan be seen through the holes which isnt element protective whatsoever but allows really good airflow. when designing the new batt i will ensure more environmental protection for the batt cells and circuitry.

the bms boards have now arivved so i can now wire it up and test the functionality of it, if this all goes well the regulator will be included and the main minimum hardware requirements of the batt will be complete and the software for battery percentage can be developed which will begin when all the currently developed electronics including the pickups are integrated and tested for any issues and to ensure everything is in working order.

the concept of the IO board has been scrapped as it was decided that it would be easier to just mount the IO jacks on a plate and wire them directly and the resistors/caps for the batt and aux in stereo to mono circuitry can be integrated into the system without any need for a dedicated board and enclosure and that a dedicated enclosure and board would take up more space and take too much time to develop for it to be worthwile when the simple and effective solution of just wiring it all together is sufficient enough.

work is being done as well now for planning and carrying out the physical construction of the body. i visited a woodshop that i plan to get the wood slab from and they said that they could use their thicknesser to make the slab 45mm thick, saving me from having to buy a large expensive thicknesser or making a router sled which both would take me more man hours. it is also decided that instead of a simple stain and  clear top coat  finish, a nice painted finish will be done as after spending about 1.5 years of my near full attention on researching, testing designing,  manuafacturing and assembling the expeditioner guitar, i should have a high quality and aestetically pleasing finish to go along with this dedication. i am planning to use a HVLP spray gun and a custom paint tent to get a good finish on the guitar, where the current plan will be to have an all black finish with a dark red 'racing strip' along the middle will be done with a matte clear top coat to achieve as close to a factory finish as possible. the planning for this is still in early stages and specific paint types will be decided soon, but i have already been researching this issue for some time and am getting there. ideally the painting plans will be completed before the body design is completed.

but for now i am loving all this progression since the start of the AFPI device design after its testing as it was the first real component that was successfuly produced since prior to that all my work was mainly theories, research and testing which took quite the toll on me. but this success has reminded me of why im actually doing this project, good things take time to happen and hard work is essential to success. so far the AFPI which includes the FX, audio, interface and RGB, and the battery are near fully operational minus a few additions that are already planned and are expected to have few difficulties. the physical body design and construction is the final section to develop before assembly of the guitar can be done and the final product can be said to be completed which is excellent considering by mid july non of these were completed, meaning in the past 2 months, nearly 2 thirds of the project have been mostly completed, with the last third already planned and have their requirements well understood and achievable.

i have about 8 weeks between now and my last exam on the 13/11/24, which was the deadline for when i had to start the main body design, which is definetely enough time to get all the existing issues squared away and im expecting to start the physical design by the 1st of November, giving me 2 months to finish the design and to manuafacture/assemble the final product. all is going to plan